+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - More efficient client reads
 - Rename server methods `prepare` for clarity
 - Improved socket state handling
+- More accurate interval time
 
 ## 0.1.0 - December 20, 2018
 First release